Solve Deepak and his journey
In this question i was making a simple task to complex task.

1. I was thinking of finding a petrol pump with minimum cost per liter and then i was trying to use priority queue and sorting to find all the stations which were coming before it.

Rather i did not understand the meaning of greedy algo.

Learning from this question is if in future i encounter a question where ith calculation depends on some previous jth value i should store the value and iterate furthur and in greedy algorithm i should think like newbie